# gen/http/openapi.yaml
swagger: "2.0"
info:
  title: Space CMS API Service
  description: Auth service for SpaceCMS. An open source content management system.
  version: ""
host: localhost:8000
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
  /api/v1/create-account:
    post:
      tags:
      - auth
      summary: CreateAccount auth
      description: Create an account by email address and password.
      operationId: auth#CreateAccount
      parameters:
      - name: CreateAccountRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/AuthCreateAccountRequestBody'
          required:
          - email
          - name
          - password
      responses:
        "201":
          description: Created response.
        "400":
          description: Bad Request response.
          schema:
            $ref: '#/definitions/AuthCreateAccountBadrequestResponseBody'
        "401":
          description: Unauthorized response.
          schema:
            $ref: '#/definitions/AuthCreateAccountUnauthorizedResponseBody'
        "404":
          description: Not Found response.
          schema:
            $ref: '#/definitions/AuthCreateAccountNotfoundResponseBody'
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/AuthCreateAccountServererrorResponseBody'
      schemes:
      - http
  /api/v1/openapi.json:
    get:
      tags:
      - auth
      summary: Download ./gen/http/openapi.json
      operationId: auth#/api/v1/openapi.json
      responses:
        "200":
          description: File downloaded
          schema:
            type: file
      schemes:
      - http
  /api/v1/tokens:
    post:
      tags:
      - auth
      summary: GrantToken auth
      description: Create an account by email address and password.
      operationId: auth#GrantToken
      parameters:
      - name: GrantTokenRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/AuthGrantTokenRequestBody'
          required:
          - email
          - password
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/AuthGrantTokenResponseBody'
            required:
            - accessToken
            - refreshToken
            - accessExpiryTime
            - refreshExpiryTime
            - accessExpiryTimeStamp
            - refreshExpiryTimeStamp
        "400":
          description: Bad Request response.
          schema:
            $ref: '#/definitions/AuthGrantTokenBadrequestResponseBody'
        "401":
          description: Unauthorized response.
          schema:
            $ref: '#/definitions/AuthGrantTokenUnauthorizedResponseBody'
        "404":
          description: Not Found response.
          schema:
            $ref: '#/definitions/AuthGrantTokenNotfoundResponseBody'
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/AuthGrantTokenServererrorResponseBody'
      schemes:
      - http
definitions:
  AuthCreateAccountBadrequestResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: true
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: false
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: true
    description: CreateAccount_badrequest_Response_Body result type (default view)
    example:
      fault: false
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: true
      timeout: true
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  AuthCreateAccountNotfoundResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: false
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: false
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: false
    description: CreateAccount_notfound_Response_Body result type (default view)
    example:
      fault: false
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: false
      timeout: false
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  AuthCreateAccountRequestBody:
    title: AuthCreateAccountRequestBody
    type: object
    properties:
      email:
        type: string
        description: Email address of the user
        example: Quia sapiente est sed accusamus temporibus.
      id:
        type: string
        description: ID of the user
        example: Laborum quam.
      name:
        type: string
        description: Full name of the user
        example: Autem rerum.
      password:
        type: string
        description: Password of the user. This is never returned to the client.
        example: Commodi fugit et voluptas.
    example:
      email: Soluta vero.
      id: Necessitatibus ipsam illum ut repellendus velit.
      name: Voluptatibus illum natus.
      password: Perferendis autem et et qui maiores non.
    required:
    - email
    - name
    - password
  AuthCreateAccountServererrorResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: true
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: false
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: false
    description: CreateAccount_servererror_Response_Body result type (default view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: false
      timeout: false
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  AuthCreateAccountUnauthorizedResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: false
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: false
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: true
    description: CreateAccount_unauthorized_Response_Body result type (default view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: false
      timeout: false
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  AuthGrantTokenBadrequestResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: false
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: true
    description: GrantToken_badrequest_Response_Body result type (default view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: true
      timeout: true
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  AuthGrantTokenNotfoundResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: false
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: true
    description: GrantToken_notfound_Response_Body result type (default view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: true
      timeout: false
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  AuthGrantTokenRequestBody:
    title: AuthGrantTokenRequestBody
    type: object
    properties:
      email:
        type: string
        description: User email
        example: Ab ea.
      password:
        type: string
        description: User password
        example: Minus possimus aut doloribus consequatur.
    example:
      email: Eos est dolor libero.
      password: Minus cupiditate adipisci omnis ex porro.
    required:
    - email
    - password
  AuthGrantTokenResponseBody:
    title: AuthGrantTokenResponseBody
    type: object
    properties:
      accessExpiryTime:
        type: integer
        example: 974932385201161190
        format: int64
      accessExpiryTimeStamp:
        type: string
        example: Repudiandae quod consectetur quia rerum.
      accessToken:
        type: string
        example: eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzaWQiOiJkYmViMjlhYTMyYjg0MTMxYTA0NjY4MDAyNzAxNWEwZSIsInJvbGUiOlsiQWRtaW5pc3RyYXRvcnMiLCJSZWdpc3RlcmVkIFVzZXJzIiwiU3Vic2NyaWJlcnMiXSwiaXNzIjoidGVzdHNpdGVjZS5sdmgubWUiLCJleHAiOjE0NTA4MzU2ODMsIm5iZiI6MTQ1MDgzMTc4M30.Yf3mmBJ8nV_IozqvvLc8L34dDklU2J7z0uXn3jsICp0
      refreshExpiryTime:
        type: integer
        example: 1419221311759041007
        format: int64
      refreshExpiryTimeStamp:
        type: string
        example: Et minus illum ducimus et.
      refreshToken:
        type: string
        example: Impedit rem eligendi consectetur voluptas rem error.
    example:
      accessExpiryTime: 3756721119481649913
      accessExpiryTimeStamp: Sit ipsum voluptatem accusamus.
      accessToken: eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzaWQiOiJkYmViMjlhYTMyYjg0MTMxYTA0NjY4MDAyNzAxNWEwZSIsInJvbGUiOlsiQWRtaW5pc3RyYXRvcnMiLCJSZWdpc3RlcmVkIFVzZXJzIiwiU3Vic2NyaWJlcnMiXSwiaXNzIjoidGVzdHNpdGVjZS5sdmgubWUiLCJleHAiOjE0NTA4MzU2ODMsIm5iZiI6MTQ1MDgzMTc4M30.Yf3mmBJ8nV_IozqvvLc8L34dDklU2J7z0uXn3jsICp0
      refreshExpiryTime: 2021119473739607656
      refreshExpiryTimeStamp: Eos recusandae quod porro.
      refreshToken: Unde iure iure nam nihil distinctio ipsum.
    required:
    - accessToken
    - refreshToken
    - accessExpiryTime
    - refreshExpiryTime
    - accessExpiryTimeStamp
    - refreshExpiryTimeStamp
  AuthGrantTokenServererrorResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: false
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: false
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: false
    description: GrantToken_servererror_Response_Body result type (default view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: false
      timeout: true
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  AuthGrantTokenUnauthorizedResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: true
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: true
    description: GrantToken_unauthorized_Response_Body result type (default view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: false
      timeout: false
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault